implement fetch and save link title
