+++ internal/infrastructure/redis.go
@@ -1,0 +1,21 @@
+package infrastructure
+
+import (
+	"os"
+
+	"github.com/go-redis/redis"
+)
+
+func NewRedisClient() (*redis.Client, error) {
+	client := redis.NewClient(&redis.Options{
+		Addr:     os.Getenv("REDIS_HOST") + ":" + os.Getenv("REDIS_PORT"),
+		Password: os.Getenv("REDIS_PASS"),
+		DB:       0,
+	})
+
+	_, err := client.Ping().Result()
+	if err != nil {
+		return nil, err
+	}
+	return client, nil
+}
